Fix for Issue "Optional loading of JQuery and JQuery-ui #13"
This fix maintains backwards compatibility of the API.

GridStackLayout works exactly as before.

Users needing to suppress the loading of JQuery and JQuery-UI should use
GridStackLayoutNoJQuery rather than GridStackLayout.

changes:
I have renamed GridStackLayout to GridStackLayoutNoJQuery and removed
the references to jquery and jquery-ui in the @JavaScript annotation

I have created a replacement GridStackLayout which extends
GridStackLayoutNoJQuery and has an @JavaScript annotation which loads
the jquery and jquery-ui as before.

All other changes are very minimal, referencing GridStackLayoutNoJQuery
where necessary.